refactor: extract ProjectWatchService from AnalyzerWrapper
